# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thapar Student Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  college_id: string;
  password: string;
}

export interface AuthUser {
  college_id: string;
  name: string;
  role: string;
}

export interface Student {
  id?: number;
  college_id: string;
  name: string;
  role?: string;
}

export interface StudentInput {
  college_id: string;
  password: string;
  name: string;
}

export interface Course {
  id: number;
  course_title: string;
  course_id: string;
  /** @nullable */
  class_nbr?: string | null;
  /** @nullable */
  subject_catalog_nbr?: string | null;
  class_section?: string;
}

export interface CourseInput {
  course_title: string;
  course_id: string;
  class_nbr?: string;
  subject_catalog_nbr?: string;
  class_section?: string;
}

export interface Assignment {
  id: number;
  course_id: number;
  assignment_name: string;
  category: string;
  /** @nullable */
  begin_date?: string | null;
  /** @nullable */
  due_date?: string | null;
  max_marks: number;
}

export interface AssignmentInput {
  course_id: number;
  assignment_name: string;
  category: string;
  begin_date?: string;
  due_date?: string;
  max_marks: number;
}

export interface Mark {
  id: number;
  student_college_id?: string;
  assignment_id: number;
  /** @nullable */
  marks_obtained: number | null;
}

export interface MarksInput {
  student_college_id: string;
  assignment_id: number;
  /** @nullable */
  marks: number | null;
}

export interface EnrollInput {
  student_college_id: string;
  course_id: number;
}

export interface AssignmentGrade {
  id: number;
  assignment_name: string;
  category: string;
  /** @nullable */
  begin_date?: string | null;
  /** @nullable */
  due_date?: string | null;
  max_marks: number;
  /** @nullable */
  marks_obtained?: number | null;
}

export interface CourseGrades {
  course: Course;
  assignments: AssignmentGrade[];
  midterm_grade: number;
  overall_grade: number;
}