# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Thapar Student Portal API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: student
    description: Student operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials

  /logout:
    get:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out

  /student/courses:
    get:
      operationId: getStudentCourses
      tags: [student]
      summary: Get enrolled courses for logged-in student
      responses:
        "200":
          description: Courses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
        "401":
          description: Unauthorized

  /student/course/{courseId}/grades:
    get:
      operationId: getStudentCourseGrades
      tags: [student]
      summary: Get assignments and grades for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course grades detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseGrades"
        "401":
          description: Unauthorized

  /admin/students:
    get:
      operationId: getAdminStudents
      tags: [admin]
      summary: List all students
      responses:
        "200":
          description: Students list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"

  /admin/student:
    post:
      operationId: createAdminStudent
      tags: [admin]
      summary: Add a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"

  /admin/student/{collegeId}/courses:
    get:
      operationId: getAdminStudentCourses
      tags: [admin]
      summary: Get courses for a student
      parameters:
        - name: collegeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"

  /admin/course:
    post:
      operationId: createAdminCourse
      tags: [admin]
      summary: Add a new course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"

  /admin/enroll:
    post:
      operationId: enrollStudent
      tags: [admin]
      summary: Enroll student in a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollInput"
      responses:
        "201":
          description: Enrolled

  /admin/assignment:
    post:
      operationId: createAdminAssignment
      tags: [admin]
      summary: Add assignment to a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"

  /admin/marks:
    post:
      operationId: updateAdminMarks
      tags: [admin]
      summary: Update marks for a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarksInput"
      responses:
        "200":
          description: Marks updated

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginInput:
      type: object
      required: [college_id, password]
      properties:
        college_id:
          type: string
        password:
          type: string

    AuthUser:
      type: object
      required: [college_id, name, role]
      properties:
        college_id:
          type: string
        name:
          type: string
        role:
          type: string

    Student:
      type: object
      required: [college_id, name]
      properties:
        id:
          type: integer
        college_id:
          type: string
        name:
          type: string
        role:
          type: string

    StudentInput:
      type: object
      required: [college_id, password, name]
      properties:
        college_id:
          type: string
        password:
          type: string
        name:
          type: string

    Course:
      type: object
      required: [id, course_title, course_id]
      properties:
        id:
          type: integer
        course_title:
          type: string
        course_id:
          type: string
        class_nbr:
          type: ["string", "null"]
        subject_catalog_nbr:
          type: ["string", "null"]
        class_section:
          type: string

    CourseInput:
      type: object
      required: [course_title, course_id]
      properties:
        course_title:
          type: string
        course_id:
          type: string
        class_nbr:
          type: string
        subject_catalog_nbr:
          type: string
        class_section:
          type: string

    Assignment:
      type: object
      required: [id, course_id, assignment_name, category, max_marks]
      properties:
        id:
          type: integer
        course_id:
          type: integer
        assignment_name:
          type: string
        category:
          type: string
        begin_date:
          type: ["string", "null"]
        due_date:
          type: ["string", "null"]
        max_marks:
          type: number

    AssignmentInput:
      type: object
      required: [course_id, assignment_name, category, max_marks]
      properties:
        course_id:
          type: integer
        assignment_name:
          type: string
        category:
          type: string
        begin_date:
          type: string
        due_date:
          type: string
        max_marks:
          type: number

    Mark:
      type: object
      required: [id, assignment_id, marks_obtained]
      properties:
        id:
          type: integer
        student_college_id:
          type: string
        assignment_id:
          type: integer
        marks_obtained:
          type: ["number", "null"]

    MarksInput:
      type: object
      required: [student_college_id, assignment_id, marks]
      properties:
        student_college_id:
          type: string
        assignment_id:
          type: integer
        marks:
          type: ["number", "null"]

    EnrollInput:
      type: object
      required: [student_college_id, course_id]
      properties:
        student_college_id:
          type: string
        course_id:
          type: integer

    AssignmentGrade:
      type: object
      required: [id, assignment_name, category, max_marks]
      properties:
        id:
          type: integer
        assignment_name:
          type: string
        category:
          type: string
        begin_date:
          type: ["string", "null"]
        due_date:
          type: ["string", "null"]
        max_marks:
          type: number
        marks_obtained:
          type: ["number", "null"]

    CourseGrades:
      type: object
      required: [course, assignments, midterm_grade, overall_grade]
      properties:
        course:
          $ref: "#/components/schemas/Course"
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/AssignmentGrade"
        midterm_grade:
          type: number
        overall_grade:
          type: number