# lib/api-zod/src/generated/types/assignment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thapar Student Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Assignment {
  id: number;
  course_id: number;
  assignment_name: string;
  category: string;
  /** @nullable */
  begin_date?: string | null;
  /** @nullable */
  due_date?: string | null;
  max_marks: number;
}