# lib/api-zod/src/generated/types/assignmentGrade.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thapar Student Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignmentGrade {
  id: number;
  assignment_name: string;
  category: string;
  /** @nullable */
  begin_date?: string | null;
  /** @nullable */
  due_date?: string | null;
  max_marks: number;
  /** @nullable */
  marks_obtained?: number | null;
}