# lib/api-zod/src/generated/types/assignmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thapar Student Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignmentInput {
  course_id: number;
  assignment_name: string;
  category: string;
  begin_date?: string;
  due_date?: string;
  max_marks: number;
}