# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thapar Student Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Course {
  id: number;
  course_title: string;
  course_id: string;
  /** @nullable */
  class_nbr?: string | null;
  /** @nullable */
  subject_catalog_nbr?: string | null;
  class_section?: string;
}