# lib/api-zod/src/generated/types/courseGrades.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thapar Student Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { AssignmentGrade } from './assignmentGrade';
import type { Course } from './course';

export interface CourseGrades {
  course: Course;
  assignments: AssignmentGrade[];
  midterm_grade: number;
  overall_grade: number;
}