# lib/api-zod/src/generated/types/courseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thapar Student Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface CourseInput {
  course_title: string;
  course_id: string;
  class_nbr?: string;
  subject_catalog_nbr?: string;
  class_section?: string;
}