# lib/api-zod/src/generated/types/enrollInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thapar Student Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface EnrollInput {
  student_college_id: string;
  course_id: number;
}