# lib/api-zod/src/generated/types/mark.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thapar Student Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Mark {
  id: number;
  student_college_id?: string;
  assignment_id: number;
  /** @nullable */
  marks_obtained: number | null;
}