# lib/api-zod/src/generated/types/marksInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thapar Student Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface MarksInput {
  student_college_id: string;
  assignment_id: number;
  /** @nullable */
  marks: number | null;
}