# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Thapar Student Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  college_id: string;
  password: string;
  name: string;
}